Retweet MapReduce Job is done.
Tweaked the location of where we write tweets.

+++ Presidential-Candidates-MapReduce/src/main/java/edu/rosehulman/csse490/mapReduce/RetweetReducer.java
@@ -18,7 +18,7 @@ public class RetweetReducer extends Reducer<Text, IntWritable, Text, DoubleWrita
 			count++;
 		}
 		
-		double average = sum / count;
+		double average = (double) sum / count;
 		
 		context.write(key, new DoubleWritable(average));
 	}
